remove separate method from macosbuilder, move method from LinuxBuilder into UnixBuilderBase
